+++ logic/state_maneger/stateManeger.cpp
@@ -9,7 +9,7 @@
 #include "../render.h"
 #include "../world.h"
 
-sf::Text makeText(const sf::Font& fontF, const std::string& text, float charSize, sf::Color color,float x, float y,const camera& cam) {
+sf::Text makeText(const sf::Font& fontF, const std::string& text, double charSize, sf::Color color,double x, double y,const camera& cam) {
     sf::Text Text;
     Text.setFont(fontF);
     Text.setString(text);
